+++ projects/angular2-smart-table/src/lib/components/cell/cell-view-mode/view-cell.component.ts
@@ -8,12 +8,16 @@ import { Cell } from '../../../lib/data-set/cell';
   template: `
     <div [ngSwitch]="cell.getColumn().type">
         <custom-view-component *ngSwitchCase="'custom'" [cell]="cell"></custom-view-component>
-        <div *ngSwitchCase="'html'" [innerHTML]="cell.getValue()"></div>
-        <div *ngSwitchDefault>{{ cell.getValue() }}</div>
+        <div *ngSwitchCase="'html'" [innerHTML]="cell.getValue()" [ngClass]="cssClass"></div>
+        <div *ngSwitchDefault [ngClass]="cssClass">{{ cell.getValue() }}</div>
     </div>
     `,
 })
 export class ViewCellComponent {
 
   @Input() cell!: Cell;
+
+  get cssClass(): string {
+    return this.cell.getColumn().classContent ?? ''
+  }
 }